[#10] Initial attempt at building for develop branch
Starting a small CI process that will build changes on push for develop branch.
Ideally, we will also want the same for PRs in this branch and I will enable this
soon. We will publish SNAPSHOT versions from this branch. And that will require
reading current version. I am yet to understand how to do that.

test: ./gradlew build local