[#9] Using checkout@v3 and pulling the full history
The version management for MapLibre-Java utility is relying on
Git tags. This eventually makes the build slower because each build
has to pull the complete git history to know the last version.

I will cut a issue to handle version control for this. For now,
I am living with pulling the full git history on build.

test: ./gradle build; dev branch deploy.