[#11] Updating the existing release process to publish to maven central
I spent a good time today to figure out where are the github secrets for this
repository. It currently does not have a package repository in Github Packages.
So, I can not publish to Github.

But I can try publishing snapshots to Maven Central. To support that I updated
the project properties to mark it with snapshot = true in this branch.

test: local gradle build;